Refactor code

+++ pkg/common/utils/utils.go
@@ -27,7 +27,7 @@ func FriendDBCopyOpenIM(dst *open_im_sdk.FriendInfo, src *db.Friend) {
 		utils.CopyStructFields(dst.FriendUser, user)
 	}
 	dst.CreateTime = uint32(src.CreateTime.Unix())
-	dst.FriendUser.CreateTime = uint32(user.CreateTime.Unix())
+	dst.FriendUser.CreateTime = user.CreateTime.Unix()
 }
 
 //
